FIX: notification service ㄱㅖ층 이름 수정
