+++ src/main/java/com/planup/planup/domain/global/message/EncouragementService.java
@@ -4,7 +4,7 @@ import com.planup.planup.domain.goal.entity.mapping.UserGoal;
 import com.planup.planup.domain.goal.service.UserGoalAggregationService;
 import com.planup.planup.domain.goal.service.UserGoalService;
 import com.planup.planup.domain.notification.dto.NotificationResponseDTO;
-import com.planup.planup.domain.notification.service.notification.NotificationServiceRead;
+import com.planup.planup.domain.notification.service.notification.NotificationQueryService;
 import com.planup.planup.domain.user.entity.User;
 import com.planup.planup.domain.user.service.query.UserQueryService;
 import org.springframework.beans.factory.annotation.Value;
@@ -27,7 +27,7 @@ public class EncouragementService {
     private final String model;
     private final UserGoalAggregationService userGoalAggregationService;
     private final UserGoalService userGoalService;
-    private final NotificationServiceRead notificationService;
+    private final NotificationQueryService notificationService;
     private final DailyLimitService dailyLimitService;
     private final UserQueryService userQueryService;
 
@@ -38,7 +38,7 @@ public class EncouragementService {
             @Value("${gemini.api-key}") String apiKey,
             UserGoalAggregationService userGoalAggregationService,
             UserGoalService userGoalService,
-            NotificationServiceRead notificationService,
+            NotificationQueryService notificationService,
             DailyLimitService dailyLimitService,
             UserQueryService userQueryService
     ) {
