+++ src/main/java/com/planup/planup/domain/notification/service/notification/NotificationCreateService.java
@@ -16,7 +16,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class NotificationCreateService {
 
-    private final NotificationServiceWrite notificationService;
+    private final NotificationCommandService notificationService;
     private final UserQueryService userQueryService;
 
     public void createChallengeNotification(Challenge challenge) {
